+++ runtime/src/swam/runtime/Instance.scala
@@ -18,6 +18,8 @@ package swam
 package runtime
 
 import formats._
+import imports._
+import exports._
 
 import internals.instance._
 import internals.interpreter._
@@ -35,15 +37,8 @@ class Instance[F[_]](val module: Module[F],
                      globals: Vector[GlobalInstance[F]],
                      private[runtime] val memories: Vector[MemoryInstance],
                      private[runtime] val funcs: Vector[CompiledFunction[F]],
-                     private[runtime] val tables: Vector[TableInstance[F]])
-    extends ImportedModule[F] {
+                     private[runtime] val tables: Vector[TableInstance[F]]) {
   self =>
-
-  def find(field: String)(implicit F: MonadError[F, Throwable]): F[ImportableInstance[F]] =
-    _exports.get(field) match {
-      case Some(f) => F.pure(f)
-      case None    => F.raiseError(new RuntimeException(s"Unknown field $field"))
-    }
 
   object exports {
     def asVar[T](name: String)(implicit F: MonadError[F, Throwable], format: ValueFormatter[T]): F[Var[F, T]] =
@@ -160,34 +155,11 @@ class Instance[F[_]](val module: Module[F],
 
 }
 
-private sealed trait ExportedField[F[_]] extends ImportableInstance[F] {
-  val tpe: Type
-}
+private sealed trait ExportedField[F[_]] extends Importable[F]
 
 private object ExportedField {
   case class Function[F[_]](tpe: FuncType, inst: CompiledFunction[F]) extends ExportedField[F]
   case class Global[F[_]](tpe: GlobalType, inst: GlobalInstance[F]) extends ExportedField[F]
   case class Table[F[_]](tpe: TableType, inst: TableInstance[F]) extends ExportedField[F]
   case class Memory[F[_]](tpe: MemType, inst: MemoryInstance) extends ExportedField[F]
-}
-
-class Val[F[_], T](global: GlobalInstance[F])(implicit F: MonadError[F, Throwable], format: ValueReader[T]) {
-
-  def apply(): F[T] =
-    format.read[F](global.value)
-
-  def value: F[T] =
-    apply()
-
-}
-
-class Var[F[_], T](global: GlobalInstance[F])(implicit F: MonadError[F, Throwable], format: ValueFormatter[T])
-    extends Val[F, T](global) {
-
-  def update(v: T): F[Unit] =
-    F.pure(global.value = format.write(v))
-
-  def :=(v: T): F[Unit] =
-    update(v)
-
 }
