+++ runtime/src/swam/runtime/Module.scala
@@ -17,6 +17,8 @@
 package swam
 package runtime
 
+import imports._
+
 import scodec.bits._
 
 import scala.language.higherKinds
@@ -39,7 +41,7 @@ class Module[F[_]](val exports: Vector[Export],
                    private[runtime] val elems: Vector[CompiledElem],
                    private[runtime] val data: Vector[CompiledData]) {
 
-  def newInstance(imports: Imports[F] = Map.empty): F[Instance[F]] =
+  def newInstance[I](imports: I)(implicit I: Imports[I, F]): F[Instance[F]] =
     engine.instantiate(this, imports)
 
   object imported {
